requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Fine-tuning libraries
peft>=0.7.0  # Parameter-Efficient Fine-Tuning (LoRA, QLoRA)
bitsandbytes>=0.41.0  # Quantization support
trl>=0.7.0  # Transformer Reinforcement Learning

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Monitoring and logging
wandb>=0.15.0
tensorboard>=2.14.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
sentencepiece>=0.1.99
protobuf>=4.24.0

# Optional: for specific models
einops>=0.7.0
scipy>=1.11.0

